Fixed README CI badge.
We were using [shields.io](https://shields.io/) to have the name of the branch showing on the travis CI badge. However, there are severe issues with this service, the badge would often fail to load. We are now back to using the normal github badge service.
Also, made the badge clickable (redirects to travis).